// parallel-web/src/types/CreateTaskRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TaskPriority } from "./TaskPriority";

export type CreateTaskRequest = { title: string, repo_ref: string, description: string, base_branch: string | null, target_branch: string | null, priority: TaskPriority | null, ssh_key_ref: string, max_execution_time: number | null, project_id: string, required_labels: { [key in string]: string }, };